pyproject.toml
[tool.poetry]
name = "sample-flask-auth"
version = "0.1.0"
description = ""
authors = ["Eduardo Pizorno <pizorno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-cors = "^5.0.0"
werkzeug = "^3.0.6"
flask-uuid = "^0.2"
flask-login = "^0.6.3"
pymysql = "^1.1.1"
cryptography = "^43.0.3"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
